[#125] Split template into multiple modules
