+++ app/Main.hs
@@ -1,6 +1,6 @@
 module Main where
 
-import System.IO (hSetEncoding, utf8)
+import System.IO (hSetEncoding, stdout, utf8)
 
 import Summoner (summon)
