+++ test/Prelude.hs
@@ -3,16 +3,7 @@
 module Prelude
        ( module Relude
        , module Relude.Extra
-
-       , endLine
-       , memptyIfFalse
        ) where
 
 import Relude
 import Relude.Extra.Enum as Relude.Extra (inverseMap, universe)
-
-endLine :: Text
-endLine = "\n"
-
-memptyIfFalse :: Monoid m => Bool -> m -> m
-memptyIfFalse p val = if p then val else mempty
